User permission related fixes
1) The first admin is not allowed to be delete / deactive or remove
admin permission.
2) Admin can view every visits, user can view his own patient’s visit.

+++ app/View/Users/index.ctp
@@ -23,16 +23,25 @@
 <?php foreach($users as $u): ?>
   <tr>
     <td><?php echo $u['User']['email'];?></td>
-    <td><?php echo $u['Profile']['firstname'] . ' ' . $u['Profile']['lastname'];?></td>
+    <td>
+      <?php echo $u['Profile']['firstname'] . ' ' . $u['Profile']['lastname'];?>
+      <?php if ($u['User']['role'] > 0): ?>
+      <span class="label label-info pull-right">
+        <span class="glyphicon glyphicon-star"></span> Admin
+      </span>
+      <?php endif; ?>
+    </td>
     <td class="actions">
       <a class="btn btn-primary" href="/users/edit/<?php echo $u['User']['id'];?>"><span class="glyphicon glyphicon-edit"></span> Edit</a>
 
-      <?php if(!$u['User']['activated']):?>
-      <a class="btn btn-default" href="/users/activate/<?php echo $u['User']['id'];?>" data-method="patch"><span class="glyphicon glyphicon-ok-circle"></span> Activate</a>
-      <?php else: ?>
-      <a class="btn btn-warning" href="/users/deactivate/<?php echo $u['User']['id'];?>" data-method="patch"><span class="glyphicon glyphicon-remove-circle"></span> Dectivate</a>
+      <?php if($u['User']['id'] != 1): ?>
+        <?php if(!$u['User']['activated']):?>
+        <a class="btn btn-default" href="/users/activate/<?php echo $u['User']['id'];?>" data-method="patch"><span class="glyphicon glyphicon-ok-circle"></span> Activate</a>
+        <?php else: ?>
+        <a class="btn btn-warning" href="/users/deactivate/<?php echo $u['User']['id'];?>" data-method="patch"><span class="glyphicon glyphicon-remove-circle"></span> Dectivate</a>
+        <?php endif; ?>
+        <a class="btn btn-danger" href="/users/delete/<?php echo $u['User']['id'];?>" data-method="delete" data-confirm="Are you sure to delete this user and all the data related to this user?"><span class="glyphicon glyphicon-trash"></span> Delete</a>
       <?php endif; ?>
-      <a class="btn btn-danger" href="/users/delete/<?php echo $u['User']['id'];?>" data-method="delete" data-confirm="Are you sure to delete this user and all the data related to this user?"><span class="glyphicon glyphicon-trash"></span> Delete</a>
     </td>
   </tr>
 <?php endforeach; ?>
